requirements.txt
fastapi==0.103.1
pydantic==1.9.2
python-dotenv==1.0.0
Requests==2.31.0
# SQLAlchemy==1.4.40
SQLAlchemy==2.0.20
uvicorn==0.22.0